requirements.txt

scikit-learn==1.7.2
crewai==0.193.2
streamlit==1.50.0
pydantic==2.11.9
pydantic-settings==2.11.0
python-dotenv==1.1.1
joblib==1.5.2
